Extracted speaker component to its own file
